+++ src/javascript/components/status.bar/status.bar.item.js
@@ -1,0 +1,26 @@
+import { puffin } from '@mkenzo_8/puffin'
+import ThemeProvider from 'ThemeProvider'
+
+const StatusBarItemBody = puffin.element(`
+    <button class="${puffin.style.css`
+        ${ThemeProvider}
+        &{
+            min-height:100%;
+            display:block;
+            margin:0;
+            padding:0 6px;
+            background:{{statusbarItemBackground}};
+            color:{{statusbarItemText}};
+            border:0px;
+        }
+        &:hover{
+            background:{{statusbarItemHoveringBackground}};
+            color:{{statusbarItemHoveringText}};
+        }
+    `}">
+
+    </button>
+`)
+
+
+export default StatusBarItemBody