+++ src/javascript/defaults/status.bar.items/cursor.position.js
@@ -1,0 +1,13 @@
+import StatusBarItem from '../../constructors/status.bar.item'
+
+const CursorPositionStatusBarItem = new StatusBarItem({
+    label:'Ln 0, Ch 0',
+    position:'left',
+    action:()=>{
+        //
+    }
+})
+
+CursorPositionStatusBarItem.hide()
+
+export default CursorPositionStatusBarItem